Finished Cooperative Multitasking Project

+++ sjkennedy43_p2/src/process.c
@@ -1,3 +1,4 @@
+#include "queue.h"
 #include "process.h"
 int create_process(int (*code_address)())// code_address is the pointer to thfuncitons that conaints the process code
 {
@@ -7,24 +8,23 @@ int create_process(int (*code_address)())// code_address is the pointer to thfun
         return -1;
     
     // set sp to stackptr + 1024 to put sp at the bottom of the stack;
-    uint32_t * sp = (uint32_t) (stackPointer + 1024);
+    
+    // set sp to stackptr + 1024 to put sp at the bottom of the stack;
+    uint64_t * sp = stackPointer + 1024;
 
     for (int i = 0; i <= 32; i++)
     {
         sp--;// decrement sp
         *sp = 0;// set uint64_t at which sp points to 0
     }
-    
-    sp[30] = (uint64_t)code_address();// [sp + 30] to code_address (type cast as uint64_t so compiler doesn't complain)
+    sp[30] = (uint64_t)code_address;// [sp + 30] to code_address (type cast as uint64_t so compiler doesn't complain)
     PCB_t * pcb = allocPCB();// pcb = allocPCB() // return a pointer back to pcb
     if (!pcb)
         return -1;// check if the pcb doesn't exist
-    pcb->sp = (uint64_t) sp;// pcb.sp = sp at the top of the stack
+    pcb->sp = sp;// pcb.sp = sp at the top of the stack
     pcb->pid = next_pid++;// pcb.pid = nextPID; // you can keep track of next_pid using next_pid =0 and next_Pid++;
 
     enqueue(&ready_queue, pcb);
 
     return 0;
-
-
 }
